Merge-fix: restore the function updateReadStatus
Eason removed the function updateReadStatus (that updates the ReadStatus of a row in a specified table) and replaced it with a less general function readMaterial (that updates the ReadStatus of a row in the 'EducationalMaterial' table).
I restored the more general function updateReadStatus, with the intention of eventually removing the readMaterial function.
The 'Read' request (which calls updateReadStatus) is used in many other places in the app, so removing it caused these other uses to fail.