Add a new function to get package contents from the DB
The previous request (and associated function) to get package contents from the database would only fetch a list of educational materials in a package (without their contents). Each package material would need to already exist in the patient's general educational material list to be displayed (the contents would be fetched from this list in the frontend). A new request and function were written to fetch package contents in full from the listener, at one level of depth. That way, the app will display all materials in a package, even if they are not duplicated in the patient's general educational material list.

The previous functionality to get package contents from the DB had been added to the function that gets educational material tables of contents from the DB. The original function was restored.