New test results module replacing existing lab results (#7)
This commit contains a few changes:

New structure for Opal requests.
- Update of the listener depedencies.
- A base class ApiRequestHandler has been created to serve as a request handler parent class
- Addition of a new validation library: express-validator, this library helps validating all the requests coming from the outside. It has been incorporated in the ApiRequestHandler.js class, for all the new requests.
- A little restructuring of the modules directory in order to fit more modules.
In the sqlInterface.js file, the sqlRunQuery function has been replaced by a new sql module under the sql folder. Now a new OpalRequest does not need to require the entire sqlInterface.js file, which comes with a lot of unrelated code just to make a query.
- All the request handlers for the new test-results modules have been put in place exemplifying the new flow and structure of Opal requests.
In terms of that last change, here are the new test result requests:

PatientTestTypes: Obtains all test types given a patient.
PatientTestTypeResults: Obtains the value results for a given test type.
PatientTestDates: Obtains an array of patient test dates.
PatientTestDateResults: Obtains the test results for a given date.
All these requests allow for obtaining lab result information as needed in the front-end, moreover, they make sense as an overall interface for patient test results. For the moment there are no limits on the queries in terms of how many we return, but it is certainly something to add later on if needed.