QSCCD-126: Only send unread notifications to the app (!103)
Modified the notification query to only return unread notifications (not all historical notifications) to the app. This is a workaround until we add pagination to the app, to cut down on the amount of data that is downloaded to the app at once.

* New behavior experienced by the app:
  - When clicking on the Notifications section, only **unread** notifications are downloaded to the app.
  - If a user clicks on a notification, it is marked as read. The notification is still available during the current session, but will be gone the next time the user logs in.
  - When a user refreshes the Notifications section, only new unread notifications are downloaded. This is essentially the same behavior as before, because new notifications are unread by design.

Note: this change shouldn't have any impact on work on the red notifications badge, which is already meant to show the number of *unread* notifications (not the total amount of historical notifications).

See merge request opalmedapps/opal-listener#103