QSCCD-205: Move registration-firebase into this project as a cron and refactor it (!115)
The [registration-firebase](https://gitlab.com/JinalVyas/registration-firebase) project has a single job, which is to clean up expired registration branches on Firebase. It was determined that this job could instead be given to the opal-listener in the form of a cron. This reduces duplicate code and the number of projects to keep track of, resulting in a cleaner architecture. This MR exists to merge `registration-firebase` into `opal-listener`.

* Moved the work done by `registration-firebase` into `opal-listener` as a cron job.
* Refactored the code to make it shorted and cleaner (async/await, etc.).
* Got rid of the stored procedures in the `registerdb` (see linked dependency) and moved the SQL into `legacy-registration/api/sql/queries.js`. After refactoring, the queries ended up being really short anyways, and this is a lot more maintainable.
* Fixed a bug where a failure in deleting one branch would cause the whole cron loop to be interrupted.

See merge request opalmedapps/opal-listener!115