QSCCD-839: Fix questionnaire compatibility with app version 1.12.2 (!157)
Closes QSCCD-839

During testing, we discovered that some requests from the version of the app currently in the stores (1.12.2) were not working with the most up-to-date listener.

This MR applies fixes to ensure that questionnaires are still working in the old version of the app. At the same time, a missing parameter was added to fix a type of notifications request.

See merge request opalmedapps/opal-listener!157