fix: remove Token parameter from request objects
This MR removes the Token parameter as a required field for request objects from the app. This parameter was never directly used or validated (only saved to the database), but its value contained sensitive data (the Firebase user's JSON Web Token), so it was decided for security reasons that this token should be removed if not used.

*Additional*
- Deleted unused files `main.spec.js` and `simulateRequest.js` which mentioned `Token`. The latter has been replaced by the files in `src/test/simulate-request`.
- Fixed one of the entries in `mock-request.js` by adding a missing line to set the request type.

See merge request opalmedapps/opal-listener!167