fix: add request to get a single questionnaire stub
Added onto the 'GetOneItem' request to be able to request a single questionnaire stub (same as those requested in 'QuestionnaireList'). This was used to fix a bug where a questionnaire could not be resumed from notifications, because some information was missing and was not being saved to the service.

*More details*

* Marked `QuestionnaireInOpalDBFromSerNum` as deprecated.
* Modified the `GetOneItem` request to be able to process single items for modules (implemented in `listener/api/modules/`).

See merge request opalmedapps/opal-listener!211