fix: reassign Dev and Staging Firebase projects
Reassigned the "opal-dev" Firebase project to Opal Dev, and created a new Firebase project for Staging. This was done both as a quick fix for the dev project (which exceeded the quota for free usage), and to improve the confusing name assignments of the Firebase projects.

See merge request opalmedapps/opal-listener#273