fix: remove registration logic and only call backend
With backend!1117 the logic in the listener is redundant and needs to be removed. This MR accomplishes that.

Simplifies the logic to verify the ID token and use `uid` and `email` in this case. The email is only sent to the backend if it is coming from a decoded token, otherwise the email needs to be taken from the email verification.

See merge request opalmedapps/opal-listener!304