refactor: restructure the project with ECMAScript module syntax (#444)
Restructured all files in this project as ECMAScript modules, which are considered the new standard for Node.js. This modernizes the project, and also unblocks updates for some dependencies which are no longer compatible with CommonJS.

- Replaced all instances of `require` with `import`.
- Replaced all instances of `module.exports` with `export` or `export default`.
- Used `with { type: "json" }` to import JSON files.
- Sorted all imports alphabetically based on eslint's `sort-imports` rule.
- Enforced file extensions in `import` statements, because the new syntax requires it, and also because it simplifies global searches for occurrences of a file.
- Deleted unused files `legacy-registration/api/utility/utility.js`, `listener/ecosystem.config.js`, and `listener/utility/queue.js`.
- Consolidated files `apiHospitalUpdate.js` and `apiPatientUpdate.js` which were no longer providing a relevant distinction between different types of requests, by just calling `sqlInterface.js` directly instead.
- Deleted some unused utility functions.

Note that this PR will be followed by another PR updating `eslint` to v9 and and reconfiguring it. Since this refactor triggers new eslint errors (which are fixed in the reconfiguration), the eslint CI job has been temporarily paused here.